requirements.txt
mysql-connector-python==8.3.0
werkzeug==3.0.1
